change to a-b lists instead of full 2-d array
to help it scale to a higher number of items and knapsack capacity,
instead of using a full 2-d array it now uses 2 rows of the array only.

For each item i, we need the array for the item i and the tem i-1 to
figure out what the solution for each weight for item i.  Because we
only need the array entries for i and i-1, we only use 2 rows of the
array instead of keeping them all.  They are called solvedA and
solvedB, and we alternate between them as we go through the list of
items.